% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{critical_values_single}
\alias{critical_values_single}
\title{Critical Values of Self-Normalization (SN) based test statistic for the
change in a single parameter (SNCP)}
\format{
A data frame with 6 variables:
\describe{
\item{\code{epsilon}}{value used to compute grid_size_scale and SN-based
test statistic}
\item{\code{0.9}}{critical value at confidence level \code{0.9}}
\item{\code{0.95}}{critical value at confidence level \code{0.95}}
\item{\code{0.99}}{critical value at confidence level \code{0.99}}
\item{\code{0.995}}{critical value at confidence level \code{0.995}}
\item{\code{0.999}}{critical value at confidence level \code{0.999}}
}
}
\usage{
critical_values_single
}
\description{
A dataset containing the critical value for SN-based change point estimates
based on the change in a single parameter.
}
\keyword{datasets}
