% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR_all.R
\name{MAR}
\alias{MAR}
\title{A funtion to generate a multivariate autoregressive process (MAR) in time
series}
\usage{
MAR(n, reptime, rho)
}
\arguments{
\item{n}{the size (length) of time series to be generated}

\item{reptime}{the number of time series to be generated}

\item{rho}{value of autocorrelation}
}
\value{
Returns a matrix of the simulated MAR processes. The number of columns
of this matrix is equivalent to the value of input argument \code{reptime}, and
the number of rows is the value of input argument \code{n}.
}
\description{
The function \code{MAR} is used for generating MAR model(s) for examples
of the functions \code{SNSeg_Uni}, \code{SNSeg_Multi}, and \code{SNSeg_HD}.
}
\examples{
MAR(n = 1000, reptime = 2, rho = -0.7)

}
